refactor(Utils): refactor helpers
