+++ src/lib/utils/defaultOptions.ts
@@ -1,4 +1,4 @@
-import type { ImperialOptions } from "../common/interfaces";
+import type { ImperialOptions } from "../types/common";
 
 export const defaultOptions: ImperialOptions = {
 	requestTimeout: 30000,
